chore : 페이지 구조 변경에 따른 파일명 및 컴포넌트명 변경
