+++ client/src/components/MindMapMainSection/ControlSection/ListView/NodeItem.tsx
