+++ client/src/components/MindMapMainSection/ControlSection/TextUpload.tsx
