+++ client/src/components/MindMapMainSection/ControlSection/UploadBox.tsx
