+++ client/src/pages/Main/index.tsx
@@ -1,32 +1,18 @@
-import Spinner from "@/components/common/Spinner";
-import Header from "@/components/Header";
-import MainSection from "@/components/MainSection";
-import Sidebar from "@/components/Sidebar";
-import NodeListProvider from "@/store/NodeListProvider";
-import { useState } from "react";
-
-import useAuth from "@/hooks/useAuth";
+import DashBoard from "@/components/Dashboard";
+import Profile from "@/components/MindMapHeader/Profile";
 
 export default function MainPage() {
-  const [isSidebarOpen, setIsSidebarOpen] = useState(true);
-
-  const toggleSidebar = () => {
-    setIsSidebarOpen(!isSidebarOpen);
-  };
-
-  const { isLoading } = useAuth();
   return (
-    <div className="flex h-full w-full">
-      {isLoading && <Spinner />}
-      <Sidebar isSidebarOpen={isSidebarOpen} toggleSidebar={toggleSidebar} />
-      <div className={`flex flex-grow flex-col transition-all duration-300 ${isSidebarOpen ? "ml-64" : "ml-0"}`}>
-        <div className="flex h-full w-full flex-col">
-          <NodeListProvider>
-            <Header />
-            <MainSection />
-          </NodeListProvider>
+    <>
+      <header className={`flex w-full justify-end bg-transparent p-4`}>
+        <Profile />
+      </header>
+      <main className="flex h-[90%] w-full flex-col overflow-hidden p-8">
+        <p className="p-3 text-2xl font-bold">대시보드</p>
+        <div className="relative flex h-[90%] w-full gap-4">
+          <DashBoard />
         </div>
-      </div>
-    </div>
+      </main>
+    </>
   );
 }
